feat: message limit report (#387)
* feat(xtask): add function to change log level at runtime

For some commands it might be desired to be less verbose as that also
reduces the resources needed. This is the case for instance of the
command to get message limits report.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor: use payload ref when building message

It saves allocations and we don't need ownership.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor: allow build_contracts to take arguments

So we can filter which contracts to build if desired. This is useful as
building contracts is slow and sometimes we might not need to build all
contracts.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* feat(xtask): cmd to generate message limit report

In order to find out what are the limit for different combinations of
numbers of signers, number of messages and message size for solana
transactions, this command performs the normal incoming flow with
different combinations of these parameters until failure.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>