refactor: do not track RotateSigners status (#391)
* refactor: do not track RotateSigners status

It's not necessary to track status of the RotateSigners command as it
doesn't go through several different states. It's either executed or
not.

With this change, we also split the GatewayExecuteData PDA as it's
different for ApproveMessages and RotateSigners, which in turns requires
us to split the initialization instructions as well.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor: improve command pda's

In order to avoid unnecessary hashing operations, the `new` method
should only be used when initializing the pda. When approving a command,
one should use the `from_bytes` function.

In order to avoid code duplication, generics was used to implement the
GatewayExecuteData PDA as it's used for both ApproveMessages and
RotateSigners commands.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(gateway): remove unnecessary check

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>