feat: flow limits (#519)
* feat: trait for dealing with rkyv PDAs

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* feat: flow limits

Add functionality to handle flow limits on interchain transfers.  As the
feature was developed, adjustments had to be made to other parts of ITS,
specially code related to role-management.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: remove outdated role-management tests

The role-management crate doesn't process the instructions anymore as
handling who has rights to deal with management is dependent on the use
case and is done outside of the crate. Thus, these test cases are
outdated.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor(flow_limit): extract and reuse

General improvements for code reuse.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix: make linter happy

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(flow_limits): wrong check logic

We want to make sure only FLOW_LIMITER role is changed and thus we
need to make sure that's the role passed in the instruction by returning
an error in case it's not.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix: remove dbg statement

This was for troubleshooting purpose and serves no use anymore.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: add comment to flow_limits test cases

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(flow-limits): add error message

This helps troubleshooting in case of errors.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>