feat(gtw): add CallContractOffchainData instruction (#568)
* feat(gtw): add CallContractOffchainData instruction

The intent is for this instruction to be used for large messages (those
too big to go through the normal ContractCall flow). The user should
call it with the payload hash only, without the actual contract call
data. The latter should be uploaded directly to the gateway together
with the signature of the event emitted by this instruction.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* test(memo): add test for call contract with offchain data

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: remove spurious println

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>