feat: its events (#68)
* feat(event-utils): extract into a crate of its own

This module was being duplicated by the gas-service and the gateway. As
we need to add events to ITS as well, it was going to be duplicated
there too. Let's reuse!

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* feat: add ITS events

Since there's a lot of events emitted by ITS, implementing serialization
and deserialization of each event was just too time consuming, and thus,
a proc-macro was introduced to makes things easier.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

wip

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(xtask): cargo tree parsing

The algorithm was messing up when it would see the `(proc-macro)`.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor(event-utils): don't box Event implementer

All of the types implementing Event so far are Sized, no need to box
anything.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(events): improve code to make clippy happy

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore: remove unused dependencies

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* docs: improve event crates doc comments

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(xtask): remove dbg logs

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>