fix(its): remove unused accounts (ACKEE-9) (#104)
* fix(its): remove unused accounts (ACKEE-9)

This was reported by Ackee during the audit:

The process_outbound_deploy oddly reads accounts from the input.
OUTBOUND_MESSAGE_ACCOUNTS_INDEX is set to 5, in case it is WithMinter
there are additional 4 accounts. In total it considers 9 , even though 2
of the accounts are not actually used. In the instruction.rs  you add
the SysvarInstruction  and the Metaplex Metadata  which are not needed.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(bindings): update after repo structure change

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): update bindings after breaking changes to API

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>