fix(its): fix LockUnlock token vulnerability (ACKEE-5) (#103)
* fix(its): fix LockUnlock token vulnerability

This issue was reported by Ackee (ACKEE-5):

There is a severe issue with the InterchainTransfer . It lacks
validation of source/destination accounts. It is possible to execute the
instruction with LockUnlock / LockUnlockFee token manager type in
unintended way -> if the attacker uses source account as the Token
Manager ATA and destination account as his ATA, it will drain the token
manager. This likely applies also to the CallContractWithInterchainToken
and CallContractWithInterchainTokenOffchainData as both uses
process_outbound_transfer.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): flaky test

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>