chore(solana): update solana-axelar dependency
As we need different solana program ids for different axelar
environments and program ids are set at compile time for solana, each id
declaration goes behind a feature flag named after the axelar
environment it's intended to be used with.

Some breaking changes to the API existed due to changes after the audit.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>