fix(gateway): confusing log message (#166)
The check does something and the message was telling something else.
Since the verification session account is created by the gateway, it can
only sign if the transaction is invoked with invoke_signed by the
gateway. We expect `initialize_payload_verification_session` to invoked
by the relayer, thus this account cannot be a signer.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>