fix: outbound deployment (#182)
* fix(its): add missing checks on remote deployment

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor(its): try different metadata sources

Instead of always relying on metaplex metadata for existing tokens, we
should try to fetch from extensions as well.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor(its): enforce token metadata for canonical tokens

This is desired according to Axelar. There's no need to require name and
symbol for custom tokens, but there's a requirement for canonical tokens
to have name and symbol.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): fallback to Metaplex metadata

In case the we cannot retrieve the metadata from the mint extension,
always fallback to try Metaplex, no matter the reason which prevented us
from getting the metadata from the mint extension.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(it): unused deps and lint

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>