chore(solana): remove unused symlinks
These symlinks were used prior to the introduction of the
`solana-axelar-cli` in order to make it easier for the CLI to locate the
JSON files with respect to the binary location instead of locating it
with respect to the directory the tool is being called from. Since the
introduction of `solana-axelar-cli`, this is not required anymore as
that script makes sure to pass the correct JSON file as argument to the
CLI program.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>