fix(its): remove unused code (#254)
* chore(its): remove unused token_manager functions

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(its): add TODO for FlowLimitSet event

We were recommended to remove this event as it's not being used. Turns
out it not being used is a mistake, we actually have to emit it.
Unfortunately we can't right now though since the current implementation
doesn't support events with Option<T>. Since we're replacing the
mechanism to emit events, it's not worth adding support for Option<T> in
the current implementation, thus adding a TODO just so we don't forget.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(its): remove unused PDA derivation functions

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(its): remove Pack from DeployApproval

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(its): remove unused dependency

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>