fix(its): decouple payer from authority accounts (#286)
* fix(its): decouple payer

Within several instructions, it's assumed the payer and the account with
some special permission/role is the same. This causes issues, specially
when the account that has the role is an account holding data, in which
case it cannot pay, causing transactions to revert.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* chore(its): some boyscouting on doc comments

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): add missing signer check to outbound transfer

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): update misleading error messages

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): properly check role transfer accounts

Origin account should always be different from destination. This issue
was reintroduced after splitting the payer from authority accounts.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* fix(its): remove unnecessary calls to `insecure_clone`

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

* refactor(program-utils): transfer funds only when necessary

If the account already has enough lamports for rent, skip transfer.

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>

---------

Signed-off-by: Guilherme Felipe da Silva <gfsilva.eng@gmail.com>